DiskSpaceManager: -fixed wrong import, fixes freebsd jail detection AdvancedSettings: -fixed edt stuff EHentaiOrg: -fixes thread 72358 PornEmbedParser: -fixes #82592 VimeoComDecrypter: -do not show broken plugin for skipped(quality) results -handle all disabled qualities FileFactory: -added method isPendingDeletion, error handling for error=105 and error=152 FilerNet: -max traffic is 125gib now OtrFilesDe: -fixed getOptionsLink method TnComAr: -fixed java compatibility UploadingSite: -reduced max connections, refs thread 72359 ClipDataCache: -updated error handling in getInternal
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@35915 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: ea24d558bfebb5521bcd741e5c3ce4c44a5698a4

+++ src/jd/plugins/hoster/TnComAr.java
@@ -126,7 +126,7 @@ public class TnComAr extends PluginForHost {
                 dllink = "https://vodgc.com/p/" + partner_id + "/sp/" + sp + "/playManifest/entryId/" + entry_id + "/flavorId/" + flavourid + "/format/url/protocol/https/a.mp4";
                 max_bitrate = max_bitrate_temp;
                 if (size != null && size instanceof Long) {
-                    filesize = (long) size;
+                    filesize = ((Long) size).longValue();
                 }
                 filesize = JavaScriptEngineFactory.toLong(entries.get("size"), 0);
             }
